+++ src/components/newModel/index.js
@@ -1,15 +1,9 @@
 import React from 'react'
 import {
-  Modal,
-  ModalOverlay,
-  ModalContent,
-  ModalHeader,
-  ModalCloseButton,
-  ModalBody,
-  ModalFooter
+  Modal
 } from '@chakra-ui/core'
 
-const NewModal = ({ mdalHeader, openModal, size, onOpen, closeButton, modalWithBody, isOpen, onClose, modalWithFooter, children, ...rest }) => {
+const NewModal = ({ mdalHeader, openModal, size, onOpen, closeButton, modalWithBody, modalcontent, isOpen, onClose, modalWithFooter, children, ...rest }) => {
   return (
     <>
       {openModal}
@@ -18,20 +12,10 @@ const NewModal = ({ mdalHeader, openModal, size, onOpen, closeButton, modalWithB
           isOpen={isOpen}
           onClose={onClose}
           size={size}
-          isCentered
+          {...rest}
         >
-          <ModalOverlay />
-          <ModalContent h={{ base: '100%', sm: 'auto' }} maxH='100%' my={{ base: 0, sm: '3.75rem' }} {...rest}>
-            {mdalHeader}
-            {!!closeButton && <ModalCloseButton />}
-            <ModalBody pb={6}>
-              {modalWithBody}
-              {children}
-            </ModalBody>
-            <ModalFooter>
-              {modalWithFooter}
-            </ModalFooter>
-          </ModalContent>
+          {modalcontent}
+          {children}
         </Modal>
       )}
     </>
