+++ stories/post.js
@@ -19,6 +19,7 @@ import MyImage3 from '../static/timg8J2T10BJ.jpg'
 import PostsPictures from '../src/components/postCard/postsPictures'
 import Model from '../src/components/newModel'
 import ModalWithHeader from '../src/components/newModel/modalWithHeader'
+import ModalConten from '../src/components/newModel/modalContent'
 import ModelButton from '../src/components/blueButton'
 
 export default {
@@ -71,9 +72,10 @@ export const nemodel = () => {
       onOpen={onOpen}
       onClose={onClose}
       w='700px'
+      modalcontent={<ModalConten
+        mdalHeader={<ModalWithHeader>標題</ModalWithHeader>}
+        modalWithFooter={<ModelButton onClick={onClose} w='100%' mt={4}>關閉</ModelButton>} />}
       openModal={<ModelButton onClick={onOpen} mt={4}>關閉</ModelButton>}
-      mdalHeader={<ModalWithHeader>標題</ModalWithHeader>}
-      modalWithFooter={<ModelButton onClick={onClose} w='100%' mt={4}>關閉</ModelButton>}
       closeButton
     // modalWithBody={<Box w='800px' bg='black' h='50px' />}
     />
